Move event display analysis task to .cxx
